Update name for an organisation moving code to controller
